+++ test/mocha/10-meters.js
@@ -1,13 +1,13 @@
 /*!
  * Copyright (c) 2021-2022 Digital Bazaar, Inc. All rights reserved.
  */
-const {AbortController} = require('abort-controller');
-const bedrock = require('bedrock');
-const {createMeter, cleanDB} = require('../helpers');
-const database = require('bedrock-mongodb');
-const {getAppIdentity} = require('bedrock-app-identity');
-const {meters} = require('bedrock-meter-usage-reporter');
-const sinon = require('sinon');
+import * as bedrock from '@bedrock/core';
+import * as database from '@bedrock/mongodb';
+import {AbortController} from 'abort-controller';
+import {createMeter, cleanDB} from '../helpers.js';
+import {getAppIdentity} from '@bedrock/app-identity';
+import {meters} from '@bedrock/meter-usage-reporter';
+import sinon from 'sinon';
 
 const meterService = `${bedrock.config.server.baseUri}/meters`;